% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poloniex.R
\name{get_from_poloniex}
\alias{get_from_poloniex}
\title{Get Data From Poloniex}
\usage{
get_from_poloniex(info_command, parameters = list(), auth_file = "")
}
\arguments{
\item{info_command}{The commmand you wish to convey. (p.s. This function is for information gathering purposes only.)}

\item{parameters}{Additional parameters to the command.}

\item{auth_file}{An authorization RDS file. See \code{\link{set_authorization}}.}
}
\description{
With this function and (optionally API credentials) you can get data from Poloniex exchange. It will return a data frame.
}
