% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poloniex.R
\name{set_authorization}
\alias{set_authorization}
\title{Create a Poloniex Credentials File}
\usage{
set_authorization(Key, Sign, file_name)
}
\arguments{
\item{Key:}{Your API key.}

\item{Sign:}{Your API sign.}

\item{file_name:}{The file name or file path you want to keep your credentials.}
}
\description{
Set up an authorization credentials file consisting of the Poloniex Key and Sign. You need to have a Poloniex account. Then access your credentials from https://poloniex.com/support/api/ (do not share your keys in any way or medium). The output is an rds file.
}
